requirements.txt
#Main Program
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
deepl==1.13.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
keyboard==0.13.5
numpy
packaging==23.0
PyAudio==0.2.13
pycparser==2.21
pydantic==1.10.5
python-dotenv==0.21.1
pywin32==305
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
SpeechRecognition==3.9.0
typing-extensions==4.5.0
urllib3==1.26.14

#MoeGoe
numba
librosa
scipy
torch
unidecode
openjtalk>=0.3.0.dev2
jamo
pypinyin
jieba
protobuf
cn2an
inflect
eng_to_ipa
ko_pron
indic_transliteration
num_thai
opencc
audonnx

#etc
faster-whisper==0.5.1
Flask==2.3.2
iso-639==0.4.5
numpy==1.24.3
#openai-whisper

#Discord
discord.py

#Web-Scrape
BeautifulSoup4
lxml
cchardet

#GUI
customtkinter
PySide6

#LoadPrompt
Pillow
pybase64
python-dateutil